Fix ILibraryAppletCreator CreateStorage, implement Write on IStorageAccessor

+++ Ryujinx.Core/OsHle/Ipc/IpcRecvListBuffDesc.cs
@@ -4,8 +4,8 @@ namespace Ryujinx.Core.OsHle.Ipc
 {
     struct IpcRecvListBuffDesc
     {
-        public long  Position { get; private set; }
-        public short Size     { get; private set; }
+        public long Position { get; private set; }
+        public long Size     { get; private set; }
 
         public IpcRecvListBuffDesc(BinaryReader Reader)
         {
@@ -13,7 +13,7 @@ namespace Ryujinx.Core.OsHle.Ipc
 
             Position = Value & 0xffffffffffff;
 
-            Size = (short)(Value >> 48);
+            Size = (ushort)(Value >> 48);
         }
     }
 }
